pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stacchip"
authors = [
  { name="Daniel Wiesmann", email="danielwiesmann@developmentseed.org" },
]
description = "Dynamically create image chips from STAC items"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pystac>=1.9.0",
    "pystac-client>=0.7.5",
    "pyarrow>=14.0.1",
    "geoarrow-pyarrow>=0.1.2",
    "rasterio>=1.3.9",
    "numpy>=1.26.0",
    "boto3>=1.29.0",
    "geopandas>=0.14.1",
    "rio-stac>=0.9.0",
    "planetary-computer>=1.0.0",
]
dynamic = [
 "version",
]

[project.urls]
"Homepage" = "https://github.com/Clay-foundation/stacchip"
"Bug Tracker" = "https://github.com/Clay-foundation/stacchip/issues"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "mock",
    "build",
]

[project.scripts]
stacchip-sentinel-1 = "stacchip.processors.sentinel_1_processor:process"
stacchip-sentinel-2 = "stacchip.processors.sentinel_2_processor:process"
stacchip-landsat = "stacchip.processors.landsat_processor:process"
stacchip-naip = "stacchip.processors.naip_processor:process"
stacchip-linz = "stacchip.processors.linz_processor:process"
stacchip-prechip = "stacchip.processors.prechip:process"
stacchip-stats = "stacchip.processors.stats:process"

[tool.hatch.version]
path = "stacchip/__init__.py"